+++ cli/src/test/java/de/gematik/refv/cli/ModuleSupportIT.java
@@ -37,7 +37,8 @@ class ModuleSupportIT {
             "isik1 src/test/resources/isik1-test.json",
             "isik2 src/test/resources/isik2-test.json",
             "isip1 src/test/resources/isip1-test.json",
-            "diga src/test/resources/diga-test.xml"
+            "diga src/test/resources/diga-test.xml",
+            "core src/test/resources/core-test.xml --profile http://fhir.org/StructureDefinition/Parameters"
     })
     @Execution(ExecutionMode.CONCURRENT)
     void testValidCliInput(String input){