+++ commons/src/test/java/de/gematik/refv/commons/validation/GenericValidatorTests.java
@@ -17,6 +17,7 @@
 package de.gematik.refv.commons.validation;
 
 import ca.uhn.fhir.context.FhirContext;
+import de.gematik.refv.commons.configuration.DependencyList;
 import de.gematik.refv.commons.configuration.ProfileConfiguration;
 import de.gematik.refv.commons.configuration.SupportedProfileVersions;
 import de.gematik.refv.commons.configuration.ValidationModuleConfiguration;
@@ -26,6 +27,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
@@ -44,6 +46,10 @@ class GenericValidatorTests {
     @SneakyThrows
     void testUnknownProfileThrowsException() {
         var configuration = new ValidationModuleConfiguration();
+        configuration.getDependencyLists().put("dependency-list", new DependencyList("1999-01-01", "2000-01-01", new ArrayList<>(), new ArrayList<>(), new ArrayList<>()));
+        configuration.getSupportedProfiles().put("https://supported-profile",new SupportedProfileVersions(new HashMap<>() {{
+            put("1.0.0", new ProfileConfiguration(List.of("dependency-list"), null));
+        }}));
         var input = "<Bundle xmlns=\"http://hl7.org/fhir\">\n"
                 + "    <id value=\"fb16b9fb-eca9-4a64-b257-083ac87c9c9c\"/>\n"
                 + "    <meta>\n"
@@ -85,5 +91,34 @@ class GenericValidatorTests {
                 + "    </meta>\n"
                 + "</Bundle>";
         Assertions.assertThrows(IllegalArgumentException.class, () -> genericValidator.validate(input, configuration));
+    }
+
+    @Test
+    @SneakyThrows
+    void testNoSupportedProfilesDefined() {
+        var configuration = new ValidationModuleConfiguration();
+        configuration.getDependencyLists().put("dependency-list", new DependencyList("1999-01-01", "2000-01-01", new ArrayList<>(), new ArrayList<>(), new ArrayList<>()));
+        var input = "<Bundle xmlns=\"http://hl7.org/fhir\">\n"
+                + "    <id value=\"fb16b9fb-eca9-4a64-b257-083ac87c9c9c\"/>\n"
+                + "    <meta>\n"
+                + "        <profile value=\"https://bla.bla|1.0.2\"/>\n"
+                + "        \n"
+                + "    </meta>\n"
+                + "</Bundle>";
+        Assertions.assertDoesNotThrow(() -> genericValidator.validate(input, configuration));
+    }
+
+    @Test
+    @SneakyThrows
+    void testNoSupportedProfilesAndNoDependencyListDefinedThrowsException() {
+        var configuration = new ValidationModuleConfiguration();
+        var input = "<Bundle xmlns=\"http://hl7.org/fhir\">\n"
+                + "    <id value=\"fb16b9fb-eca9-4a64-b257-083ac87c9c9c\"/>\n"
+                + "    <meta>\n"
+                + "        <profile value=\"https://bla.bla|1.0.2\"/>\n"
+                + "        \n"
+                + "    </meta>\n"
+                + "</Bundle>";
+        Assertions.assertThrows(IllegalStateException.class, () -> genericValidator.validate(input, configuration));
     }
 }